+++ ItHappened.Infrastructure/EFCoreRepositories/EFEventsRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using AutoMapper;
 using ItHappened.Domain;
-using ItHappened.Infrastructure.Mappers;
+using ItHappened.Infrastructure.Dto;
 using LanguageExt.UnsafeValueAccess;
 
 namespace ItHappened.Infrastructure.EFCoreRepositories
